requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
langchain-ollama
pypdf
bs4 
arxiv
pymupdf
langchain-text-splitters
langchain-openai
chromadb
faiss-cpu
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
langgraph
langgraph-cli[inmem]